requirements.txt
django==5.2.12
djangorestframework==3.17.0
django-cors-headers==4.9.0
drf-spectacular==0.28.0
pytest==8.3.5
pytest-django==4.9.0